modularize package function
Splintering the installation directives for optional parts of extensions
like schemas saves guarding them with if blocks and keeps the PKGBUILDs
free from unneeded parts. This is especially important to consider as
the PKGBUILDs are uploaded to the AUR in their composed form for
everyone to read, so striving for an at least somewhat readable format
there is worthwhile.

Unfortunately, this makes including the templates a little more
cumbersome. It would have been possible to shorten this up a bit by
recursively stating templates in other templates, but this would, again,
make the resulting PKGBUILD very cluttered.

Note that the package functions share a common namespace. That is why
they are ordered by numbers so that a later one may use a variable
defined by a previous component