parse metadata for name
Many developers do not place the extension code into a directory named
equal to the extension. The metadata.json file on the other hand always
has to be placed in the extensions root directory. Telling name and
location from it enables automatic installation out of arbitrary or even
no directories.