also install supplementary files
Some extensions also come with style sheets and interface definitions.
Find catches them all.

Replacing a glob with a find also has the effect that the process still
completes successfully when nothing is found. This is somewhat averse
because now packaging will not fail any longer if no code is installed.
On the other hand, this is coherent to the installation of schemas.