switch to branches using convenience.js
The convenience.js taken from GNOME's own extensions is part of most of
the packages. It cares for looking up schemas and locales from the
adequate locations and permits global installations. Some extensions
lacked this feature. They were adapted and the changes were filed as
pull requests. Until they are merged, the forks are used herein.