unify patching package versions
These are difficult to export to a template because the versions to add
may differ for each package and templates cannot be parameterized.
Setting a variable in the script and reading it in the template would be
possible, but would clutter the resulting PKGBUILD. Keeping these
readable is more important, as they show up in the AUR.

What can be done, nonetheless, is always keeping the change in prepare
so that build stays free for later use.