limit higher versions
Restricting to greater or equal version 3.14 will not permit
installation on 3.14.1 or any further minor update to 3.14, though
compatibility is normally not broken with these. The real break happens
when the version number changes to a new (odd) development release
number, so restricting the version to be less than the development
number catches all 3.14.x numbers.