bump all versions
Now that templates are sufficiently implemented, they bring benefits or
changes to all packages, so an update for all of them is appropriate.

The pkgrel of all PKGBUILDs that did not increase their pkgver
automatically was incremented, so that all of them will update.

A few packages are exceptional, as they switched their versioning scheme
from revision-count to tags, which indeed decreased the pkgver. Getting
them to update now nonetheless could be achieved setting an epoch of 1.
With gnome-shell release version 3.18 at the latest, they will report to
be incompatible and cease to work, forcing a manual update anyways.
Considering this, epoch was left alone as, "epoch should only be used
when absolutely required to do so" as of
https://wiki.archlinux.org/index.php/PKGBUILD.