install using templates
The Makefile includes an target "_build" that composes all files ready
for installation of the same name intended as an intermediate step
before locally installing or archive packaging. Exploiting this target
leaves a state ready for our templates to install.

The templates locate the extension base directory by the "metadata.json"
file residing in it. After the build step, there are now two such files
in the $srcdir, the original source one and a processed variant in the
_build directory. Due to natural alphabetic sorting by find, the latter
is fortunately found first.  This is exactly the one to use, so any
further results are now discarded not to confuse cd.