simplify distribution
Now that no package is marked as planned to be replaced any longer,
there are no files to ignore. So distribute.sh can be replaced with a
simple call to makepkg-expanded with its new -b flag.

This essentially reverts the last commit to "provide a distribution
helper" apart from the change to .gitignore, as it is not the
distribution helper but makepkg-expanded that leaves the files to be
ignored.