fix version comparison
The gnome-shell-version template only yields the minor version.
Comparing it to a string of minor and major will never succeed.

Originally, a restriction on the current shell version was omitted for
historical reasons - most packages did so when dependencies were defined
manually. This now turns into a concious decision to

- permit the usage of the extension in instable gnome-shell versions -
  the ones with odd minor version numbers - where incompatibilities are
  introduced gradually.

- ease the transition from one stable gnome-shell version - the ones
  with even minor version numbers - to the next. In an ideal world, all
  extensions should have been updates before the update hits the
  repositories. But as Arch is quite more on the edge than developers of
  some of the extensions, this is not the case more often than not.

Note that the analogue package collection of Mozilla extensions [1],
faces a problem similar to the second point. In contrast, restrictions
are not alleviated there because, addons.mozilla.org is queried about
the compatibilities and usually answers positively. Firefox goes through
a long release process of an nightly and beta releases with an active
user base, so Add-Ons are usually declared compatible to future versions
by default as long as experience does not show otherwise.

[1]: https://github.com/dffischer/mozilla-extensions