improve dependency when lacking sources
To mitigate the consequences of generating the .SRCINFO without sources
present, at least gnome-shell without any version will always be noted
as a dependency.

This improves contribution and AUR version 4 compatibility, as a
reasonably informative .SRCINFO can now be generated in a separate
repository without the need to download the sources there.