do not generate version range until packaging
Before, the dependency on gnome-shell could easily end up inconsistently
version restricted or plain, depending on whether the sources were
already downloaded before building every single package.

This commit fixes this by always placing a plain dependency for
mksrcinfo to find. When packaging, it is replaced by a range. This is
done in a package subfunction, hidden from mksrcinfo.