automatically determine last version
Contrary to activities-config, this is not a matter of simply parsing
the respective extensions.gnome.org page. This extension releases
different variants for different versions of the gnome-shell. The
variant for the latest shell version does not always have the highest
version number. Thus it is easier to just parse the download page from
the extension's own website, which always lists these first.