improve location metric
Some extension repositories include directories with templates or
partial variants of the extension, which may be falsely located when
searching only for the metadata.json. As an extension consists of both
the metadata description and a JavaScript entry point always named
extension.js, searching for both only matches final composed extension.