use original image
Meanwhile, the bug in librsvg propagated upstream. [1]

Though it has not been fixed as of now, the maintainer of the extension
cared for the issue as well [2] and was so kind to rework the image not
to trigger it to begin with [3]. So there is no need to patch the image
any longer and the respective commit "patch up unreadable image" can be
reverted, bringing this package as close to upstream as it has been.

This change does not increase the pkgrel as there is no functional
change to be shipped out. Already installed packages will continue to
work without the need of an update.

[1]: https://bugzilla.gnome.org/show_bug.cgi?id=759084
[2]: https://github.com/micheleg/dash-to-dock/issues/262
[3]: https://github.com/micheleg/dash-to-dock/commit/8a7e74e