re-adopt dash-to-dock
This reverts "hand off dash-to-dock".

After the recent conversion of the logo.svg to work around a bug in
librsvg, carlgeorge transferred ownership back.

Both the work in this repository and his interim maintenance started
from the same base, left behind by a former maintainer unknown to me by
now. It can still be found in the commit "import dash-to-dock".
Improvements made by carlgeorge will be ported into dash-to-dock or
dash-to-dock-git as following commits.