propagate upstream update
Apart from translation updates, the intellihide machinery has been
optimized. The extension can now also be active with full screen windows
and supports the upcoming gnome-shell version 3.20.

Fortunately, the same dash-to-dock release applies to gnome-shell 3.18
and 3.20 as well. This saves the hassle of rewriting the code that
extracts the download link from the webpage for the non-git variant and
provides the hopeful perspective that transition will be effortless when
the next stable shell version hits the Arch Linux repositories.