choose newest extension source
To always find the correct location to install from, built extensions
have to be told apart from their templates or components. The find
instruction employed before this commit relied on that the extension.js
was added in the built process, which did not prove to be a reliable
metric especially with the Git variant of dash-to-dock.

Thus, the first idea to develop a more reliable location strategy was to
add a test on whether no Makefile exists in the directory in question.
This would solve the case of dash-to-dock, but break extensions like
hidetopbar, which use a Makefile more for purposes of packing a Zip
file. These Makefiles for distribution only are not used to build the
project, which is why the package just starts to install from the
directory containing the Makefile. But this location would be rejected
by a process that decides upon the absence thereof.

So, all possible locations are collected and sorted by their
modification time. As a build process happens after unpacking of the
sources, a built variant always sorts first by this metric. Extensions
that are simply installed without any build step are unaffected as there
is only one location in question.

This does not help to tell apart partial components, like present in
CoverflowAltTab, where a second directory containing only a
metadata.json is present to aid packaging. To not have the find call be
accidentally trapped in such a directory, the condition that an
extension must be minimally made up of a metadata,json and an
extension.js is kept in place.