locally install and compile schema
An intermediary commit removed the ability to find the schema when it is
installed in the global location /usr/share/glib-2.0/schemas. Now, only
a schema subdirectory in the extension root is examined.

This may be an overeager side effect: The respective commit comment
talks about fixing translations instead, which continue to be loaded
from the global location /usr/share/locale.

Nonetheless, the schema has to be installed and compiled inside the
extension directory now, which is what the install directives normally
do. Only the post-installations script now has to be patched to compile
the installed schema despite a DESTDIR set to $pkgdir.