remove erroneous tag
Version 3.4 was accidentally tagged with a comma, which the pkgver
function may pick up while querying the history for a version number.
The comma could introduce complications to version complication, leading
to unnoticed updates in the future. Removing the tag ensures proper
versioning.

The package contents are unchanged by this, so the pkgrel remains
constant.